image_processing_pipeline/requirements.txt
kafka-python
pymongo
pillow